+++ iot_temp_humi/inc/get_time.h
@@ -1,0 +1,18 @@
+/*** 
+ * @Author: RoxyKko
+ * @Date: 2023-04-04 17:04:22
+ * @LastEditors: RoxyKko
+ * @LastEditTime: 2023-04-04 17:26:19
+ * @Description: 获取时间函数
+ */
+
+#ifndef __GET_TIME_H__
+#define __GET_TIME_H__
+
+#include "iot_main.h"
+#include <sys/time.h>
+
+double get_time(char *datime);
+
+
+# endif
